test(dcode): format and split managed patch contract
